requirements.txt
pysam
pyyaml
# scipy
# seaborn
numpy
matplotlib
reportlab
# pyspark